py/plugins/vertex-ai/pyproject.toml
[project]
authors = [{ name = "Google" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Environment :: Web Environment",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries",
]
dependencies = ["genkit", "google-cloud-aiplatform>=1.77.0", "pytest-mock"]
description = "Genkit Google Cloud Vertex AI Plugin"
license = { text = "Apache-2.0" }
name = "genkit-vertex-ai-plugin"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires      = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/genkit", "src/genkit/plugins"]
